Render targets, removing Draw logic

+++ src/Atom/Engine/Graphics/Image/Formats/D/Depth.cs
@@ -17,21 +17,21 @@ public static class Depth
     [ImageFormat(ImageFormat.D16_UNorm)]
     public struct D16_UNorm : IImageFormat
     {
-        public ushort D;
+        public f16 D;
     }
     
     [ImageFormat(ImageFormat.D32_SFloat)]
     public struct D32_SFloat : IImageFormat
     {
-        public float D;
+        public f32 D;
     }
     
     [StructLayout(LayoutKind.Sequential, Pack = 4)]
     [ImageFormat(ImageFormat.D16_UNorm_S8_UInt)]
     public struct D16_UNorm_S8_UInt : IImageFormat
     {
-        public Half D;
-        public byte S;
+        public f16 D;
+        public u8  S;
     }
     
     [StructLayout(LayoutKind.Sequential, Pack = 8)]
@@ -41,15 +41,15 @@ public static class Depth
         // Special behavior on copy: 
         // on buffer copy, it is possible to select either D or S
         // D is packed in 32 bit, the first 24bit being the data and 8 last being undefined.
-        public float D; // this is utterly wrong eh, but at least the S is placed correctly... lol
-        public byte S;
+        public f32 D; // this is utterly wrong eh, but at least the S is placed correctly... lol
+        public u8  S;
     }
     
     [StructLayout(LayoutKind.Sequential, Pack = 8)]
     [ImageFormat(ImageFormat.D32_SFloat_S8_UInt)]
     public struct D32_SFloat_S8_UInt : IImageFormat
     {
-        public float D;
-        public byte S;
+        public f32 D;
+        public u8  S;
     }
 }
